+++ packages/cysync-core/src/constants/i18n/en.ts
@@ -555,6 +555,7 @@ const en = {
         toggle: 'Send Max',
         dollar: '$',
         error: 'Insufficient funds',
+        notOverDustThreshold: 'Amount is lower than dust limit',
       },
       fees: {
         title: 'Fees',