+++ packages/db-interfaces/src/entities/index.ts
@@ -7,3 +7,4 @@ export * from './transaction';
 export * from './wallet';
 export * from './transactionNotificationRead';
 export * from './transactionNotificationClick';
+export * from './migrations';