+++ apps/desktop/extraResources/RELEASE_NOTES.md
@@ -1,3 +1,13 @@
+# v2.0.1
+
+### 🚀 Features
+
+- Added more ERC-20 token for evm chains
+
+### 🐛 Bugfixes
+
+- Various minor bug fixes
+
 # v2.0.0
 
 ### 🚀 Features