+++ packages/coin-support-interfaces/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cypherock/coin-support-interfaces
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`1bdcfb1`](https://github.com/Cypherock/cypherock-cysync/commit/1bdcfb1c02a7369ae0810ca00bf19af42c65bef7)]:
+  - @cypherock/db-interfaces@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes