+++ packages/db-interfaces/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cypherock/db-interfaces
 
+## 0.0.2
+
+### Patch Changes
+
+- [#362](https://github.com/Cypherock/cypherock-cysync/pull/362) [`1bdcfb1`](https://github.com/Cypherock/cypherock-cysync/commit/1bdcfb1c02a7369ae0810ca00bf19af42c65bef7) Thanks [@irshadCypherock](https://github.com/irshadCypherock)! - updated erc20 token list for evm chains
+
 ## 0.0.1
 
 ### Patch Changes